+++ concrete_datastore/api/v1/views.py
@@ -1811,7 +1811,7 @@ class ApiModelViewSet(PaginatedViewSet, viewsets.ModelViewSet):
         def check_date_format(date_type, param_values):
             date_format = 'yyyy-mm-dd'
             date_regex = r'^\d{4}-\d{2}-\d{2}$'
-            date_time_format = 'yyyy-mm-ddThh:mm:ss.xxxxxxZ'
+            date_time_format = 'yyyy-mm-ddThh:mm:ss[.xxxxxx]Z'
             date_time_regex = r'^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(\.\d{1,6})?Z$'
 
 
@@ -1843,7 +1843,7 @@ class ApiModelViewSet(PaginatedViewSet, viewsets.ModelViewSet):
                         False,
                         Response(
                             data={
-                                'message': "Wrong date format, should be '{}' or 'yyyy-mm-ddThh:mm:ss' or '{}'".format(
+                                'message': "Wrong date format, should be '{}' or '{}'".format(
                                     date_format, date_time_format
                                 ),
                                 '_errors': ['INVALID_QUERY'],